requirements.txt
transformers==4.46.3
python-dotenv==1.0.0
accelerate>=0.26.0
bitsandbytes>0.37.0
trl==0.12.1
peft==0.13.2
--find-links https://download.pytorch.org/whl/cu124
torch
tqdm
scikit-learn
matplotlib
nltk